+++ ldap2pg/psql.py
@@ -1,10 +1,11 @@
 from __future__ import unicode_literals
 
 import logging
+import re
 
 import psycopg2.extensions
 
-from .utils import AllDatabases, UserError
+from .utils import AllDatabases, UserError, urlparse, urlunparse
 
 
 logger = logging.getLogger(__name__)
@@ -12,6 +13,22 @@ logger = logging.getLogger(__name__)
 
 psycopg2.extensions.register_type(psycopg2.extensions.UNICODE)
 psycopg2.extensions.register_type(psycopg2.extensions.UNICODEARRAY)
+
+
+_dbname_re = re.compile("dbname *= *'?[^ ]*'?")
+
+
+def inject_database_in_connstring(connstring, dbname):
+    if dbname is None:
+        return connstring
+
+    if connstring.startswith('postgres://'):
+        pr = list(urlparse(connstring))
+        pr[2] = dbname
+        return urlunparse(pr)
+    else:
+        connstring = _dbname_re.sub('', connstring)
+        return connstring + " dbname=%s" % (dbname,)
 
 
 class PSQL(object):
@@ -24,25 +41,19 @@ class PSQL(object):
         self.pool = {}
         self.max_pool_size = max_pool_size
 
-    def __call__(self, dbname):
-        if dbname not in self.pool and len(self.pool) >= self.max_pool_size:
+    def __call__(self, dbname=None):
+        if dbname in self.pool:
+            session = self.pool[dbname]
+        elif len(self.pool) >= self.max_pool_size:
             msg = (
                 "Database limit exceeded.\n"
                 "ldap2pg doesn't support cluster with more than %d databases."
             ) % (self.max_pool_size)
             raise UserError(msg)
-
-        if self.connstring.startswith('postgres://'):
-            if '/?' in self.connstring:
-                connstring = self.connstring.replace('/?', '/%s?' % (dbname,))
-            elif '?' in self.connstring:
-                connstring = self.connstring.replace('?', '/%s?' % (dbname,))
-            else:
-                connstring = self.connstring.rstrip('/') + "/%s" % (dbname,)
         else:
-            connstring = self.connstring + " dbname=%s" % (dbname,)
+            connstring = inject_database_in_connstring(self.connstring, dbname)
+            self.pool[dbname] = session = PSQLSession(connstring.strip())
 
-        session = self.pool.setdefault(dbname, PSQLSession(connstring.strip()))
         return session
 
     def itersessions(self, databases):
@@ -61,19 +72,19 @@ class PSQLSession(object):
 
     def __del__(self):
         if self.cursor:
-            logger.debug("Closing Postgres cursor to %s.", self.connstring)
             self.cursor.close()
             self.cursor = None
         if self.conn:
-            logger.debug("Closing Postgres connexion to %s.", self.connstring)
+            logger.debug(
+                "Closing Postgres connexion to '%s'.", self.connstring)
             self.conn.close()
             self.conn = None
 
     def __enter__(self):
         if self.conn:
-            logger.debug("Using Postgres connection to %s.", self.connstring)
+            logger.debug("Using Postgres connection to '%s'.", self.connstring)
         else:
-            logger.debug("Connecting to Postgres %s.", self.connstring)
+            logger.debug("Connecting to Postgres '%s'.", self.connstring)
             self.conn = psycopg2.connect(self.connstring)
         if not self.cursor:
             self.cursor = self.conn.cursor()