+++ ldap2pg/role.py
@@ -41,7 +41,7 @@ class Role(object):
     def create(self):
         yield Query(
             'Create %s.' % (self.name,),
-            'postgres',
+            None,
             dedent("""\
             CREATE ROLE "{role}" WITH {options};
             COMMENT ON ROLE "{role}" IS 'Managed by ldap2pg.';
@@ -50,7 +50,7 @@ class Role(object):
         if self.members:
             yield Query(
                 'Add %s members.' % (self.name,),
-                'postgres',
+                None,
                 'GRANT "%(role)s" TO %(members)s;' % dict(
                     members=", ".join(map(lambda x: '"%s"' % x, self.members)),
                     role=self.name,
@@ -63,7 +63,7 @@ class Role(object):
         if self.options != other.options:
             yield Query(
                 'Update options of %s.' % (self.name,),
-                'postgres',
+                None,
                 dedent("""\
                 ALTER ROLE "{role}" WITH {options};
                 COMMENT ON ROLE "{role}" IS 'Managed by ldap2pg.';
@@ -79,7 +79,7 @@ class Role(object):
                 )
                 yield Query(
                     'Add missing %s members.' % (self.name,),
-                    'postgres',
+                    None,
                     "GRANT \"%(role)s\" TO %(members)s;" % dict(
                         members=", ".join(map(lambda x: '"%s"' % x, missing)),
                         role=self.name,
@@ -89,7 +89,7 @@ class Role(object):
             if spurious:
                 yield Query(
                     'Delete spurious %s members.' % (self.name,),
-                    'postgres',
+                    None,
                     "REVOKE \"%(role)s\" FROM %(members)s;" % dict(
                         members=", ".join(map(lambda x: '"%s"' % x, spurious)),
                         role=self.name,
@@ -109,7 +109,7 @@ class Role(object):
         )
         yield Query(
             'Drop %s.' % (self.name,),
-            'postgres',
+            None,
             "DROP ROLE \"%(role)s\";" % dict(role=self.name),
         )