+++ docs/changelog.md
@@ -11,6 +11,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 - Lint log level and messages.
 - Deduplicate LDAP auto-attributes.
+- Add `parents_attribute` to fetch parent from LDAP entry.
 
 
 # ldap2pg 4.4