+++ docs/changelog.md
@@ -12,6 +12,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Lint log level and messages.
 - Deduplicate LDAP auto-attributes.
 - Add `parents_attribute` to fetch parent from LDAP entry.
+- Comment roles with `Managed by ldap2pg.`.
 
 
 # ldap2pg 4.4