+++ docs/changelog.md
@@ -11,6 +11,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 - Allow to inspect owners per schema.
 - Use configured database instead of hardcoded `postgres`.
+- Increase arbitrary database limit to 256.
 
 
 # ldap2pg 4.5